--- lib/stretchy/refreshable.rb
@@ -1,15 +1,0 @@
-module Stretchy
-  module Refreshable
-    extend ActiveSupport::Concern
-
-    included do
-      after_save :refresh_index
-      after_destroy :refresh_index
-    end
-
-    def refresh_index
-      self.class.refresh_index!
-    end
-
-  end
-end
