+++ lib/stretchy/relations/finder_methods.rb
@@ -3,6 +3,8 @@ module Stretchy
 
       module FinderMethods
 
+        METHODS = [:first, :first!, :last, :last!]
+
         def first
           return results.first if @loaded
           spawn.first!.results.first
@@ -39,24 +41,6 @@ module Stretchy
           self
         end
 
-        # size is not permitted to the count API but queries are.
-        #
-        # if a query is supplied with `.count` then the user wants a count of the results
-        # matching the query
-        #  
-        # however, the default_size is used to limit the number of results returned
-        # so we remove the size from the query and then call `.count` API
-        #
-        # but if the user supplies a limit, then we should assume they want a count of the results
-        # which could lead to some confusion 
-        # suppose the user calls `.size(100).count` and the default_size is 100
-        # if we remove size from the query, then the count could be greater than 100
-        # 
-        # I think the best way to handle this is to remove the size from the query only if it was  
-        # applied by the default_size
-        # If the user supplies a limit, then we should assume they want a count of the results
-        # 
-        # if size is called with a limit,
         def count
           return results.count if @loaded || @values[:size].present?
           spawn.count!
